Cleaned-up code for cleaning up uncompiled .py files.

+++ dyndns/dyndns.py
@@ -1154,38 +1154,21 @@ LOG_FILE_0, LOG_FILE_FULL_PATH, LOG_FILE_CURATED_PATH, VID_TO_CURATE = LogFileIn
 SUBFOLDER = "./params"
 
 PARAMSFILES = [f for f in listdir(SUBFOLDER) if isfile(join(SUBFOLDER, f))]
-
-print(PARAMSFILES)
-
-print(SUBFOLDER)
-SUBFOLDER = "./params"
 SUBFOLDER = SUBFOLDER.replace('./','')
-print(SUBFOLDER)
-
-
-
 FULL_SCRIPT_PATH = os.path.dirname(os.path.abspath(__file__))
 FULL_SUBFOLDER_PATH = join(FULL_SCRIPT_PATH, SUBFOLDER)
-print(join(FULL_SCRIPT_PATH, SUBFOLDER))
-print(FULL_SUBFOLDER_PATH)
 
-# First confirm a .pyc copy of each file exists.
-for filename in PARAMSFILES:
-	print(filename)
-	if '__init__' not in filename and '_TEMPLATE' not in filename:
-		if '.pyc' in filename:
-			print(".pyc file detected = "+filename)
-			PY_FILE = filename.replace('.pyc','.py')
-			# For each .pyc file, check if a corresponding .py file exists
-			for uncompfile in PARAMSFILES:
-				if (uncompfile == PY_FILE):
-					print("CLEAN_UNCOMPILED = Deleting "+uncompfile)
-					os.remove(join(FULL_SUBFOLDER_PATH, uncompfile))
-
-
-
-
-
+if CLEAN_UNCOMPILED:
+	# First confirm a .pyc copy of each file exists.
+	for filename in PARAMSFILES:
+		if '__init__' not in filename and '_TEMPLATE' not in filename:
+			if '.pyc' in filename:
+				# For each .pyc file, check if a corresponding .py file exists
+				PY_FILE = filename.replace('.pyc','.py')
+				for uncompfile in PARAMSFILES:
+					if (uncompfile == PY_FILE):
+						print("CLEAN_UNCOMPILED = Deleting "+uncompfile)
+						os.remove(join(FULL_SUBFOLDER_PATH, uncompfile))
 
 # 2. Get current public IP --------------------------------------------------------------------------------
 # ---------------------------------------------------------------------------------------------------------
